@@ -3,7 +3,7 @@
     "description": "Validation resources and rules.",
     "type": "project",
     "require": {
-        "php": ">=7.3",
+        "php": ">=7.4.2",
         "ext-json": "*",
         "ext-mbstring": "*",
         "jitesoft/exceptions": "2.*"
@@ -26,7 +26,7 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.1.5",
+        "phpunit/phpunit": "^9.2",
         "fzaninotto/faker": "^1.9",
         "squizlabs/php_codesniffer": "^3.5"
     },
